빌드업계산기 초안 개발

+++ src/main/java/com/rocky/appstockdata/domain/BuildUp.java
@@ -7,12 +7,12 @@ import lombok.Setter;
 @Getter
 @Setter
 public class BuildUp {
-    private Float earningRate;
+    private Double earningRate;
     private Long earningAmount;
     private Long totalAmount;
 
     @Builder
-    public BuildUp(Float earningRate, Long earningAmount, Long totalAmount) {
+    public BuildUp(Double earningRate, Long earningAmount, Long totalAmount) {
         this.earningRate = earningRate;
         this.earningAmount = earningAmount;
         this.totalAmount = totalAmount;
